system76/common: Add battery abstraction
