+++ src/board/system76/common/common.mk
@@ -24,6 +24,10 @@ SRC+=$(wildcard $(SYSTEM76_COMMON_DIR)/*.c)
 INCLUDE+=$(wildcard $(SYSTEM76_COMMON_DIR)/include/common/*.h) $(SYSTEM76_COMMON_DIR)/common.mk
 CFLAGS+=-I$(SYSTEM76_COMMON_DIR)/include
 
+# Add charger
+CHARGER?=bq24780s
+SRC+=$(SYSTEM76_COMMON_DIR)/charger/$(CHARGER).c
+
 # Add scratch ROM
 include $(SYSTEM76_COMMON_DIR)/scratch/scratch.mk
